+++ src/main/java/com/zagurskaya/cash/model/dao/RateCBDao.java
@@ -4,6 +4,7 @@ import com.zagurskaya.cash.entity.RateCB;
 import com.zagurskaya.cash.exception.DaoException;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface RateCBDao extends Dao<RateCB> {
@@ -16,5 +17,5 @@ public interface RateCBDao extends Dao<RateCB> {
      * @return true on successful creation
      * @throws DaoException database access error or other errors
      */
-    Double rateCBToday(LocalDateTime now, Long coming, Long spending) throws DaoException;
+    BigDecimal rateCBToday(LocalDateTime now, Long coming, Long spending) throws DaoException;
 }
